backend/requirements.txt
fastapi
uvicorn
sentence-transformers
chromadb
python-dotenv
pydantic
python-multipart
langchain
groq
tiktoken
g4f==0.1.7.0
